modularize jobs
