--- app/jobs/message_sender_job.rb
@@ -1,5 +1,0 @@
-class MessageSenderJob < ApplicationJob
-  def perform(message_id)
-    MessagingService.send Message.eager_load(:volunteer, :creator, request: :organisation).find(message_id)
-  end
-end
